Update the Wiki Loves Monument entry (#21)
* campagins: updae WLM link

The old link seems to contain information about the 2019 edition of WLM. It
doesn't seem to be containing any information about the 2020 edition.

So, use a link that contains information about the 2020 edition.

* campaigns: clarify that WLM is not just about Wikipedia

WLM is held by many Wikimedia chapters. So, it's misleading to call it
a "Wikipedia photo contest". Correct it by using a more general description
which also clarifies that the event focuses on monuments.